+++ tests/os_open_test.py
@@ -93,8 +93,7 @@ class TestSecureCodingStandardChecker(pylint.testutils.CheckerTestCase):
             """
         )
 
-        self.checker.set_os_open_mode('True')
-        assert self.checker._prefer_os_open
+        self.checker.set_os_open_allowed_modes('True')
 
         call_nodes = []
         with_nodes = []
@@ -123,10 +122,16 @@ class TestSecureCodingStandardChecker(pylint.testutils.CheckerTestCase):
         ),
     )
     def test_os_open_call_default_modes(self, mode, arg, expected_warning):
-        node = astroid.extract_node(f'os.open("file.txt", os.O_WRONLY, 0o{mode:o}) #@')
-        self.checker.set_os_open_mode(arg)
+        code = f'os.open("file.txt", os.O_WRONLY, 0o{mode:o}) #@'
+        print(code)
+        node = astroid.extract_node(code)
+        self.checker.set_os_open_allowed_modes(arg)
         if expected_warning:
-            with self.assertAddsMessages(pylint.testutils.Message(msg_id='os-open-unsafe-permissions', node=node)):
+            with self.assertAddsMessages(
+                pylint.testutils.Message(
+                    msg_id='os-open-unsafe-permissions', node=node, args=(getattr(self.checker, '_os_open_msg_arg'),)
+                )
+            ):
                 self.checker.visit_call(node)
         else:
             with self.assertNoMessages():
@@ -153,9 +158,13 @@ class TestSecureCodingStandardChecker(pylint.testutils.CheckerTestCase):
     )
     def test_os_open_call(self, mode, call_mode, arg, expected_warning):
         node = astroid.extract_node(call_mode.format(mode))
-        self.checker.set_os_open_mode(arg)
+        self.checker.set_os_open_allowed_modes(arg)
         if expected_warning and mode != 0o755:
-            with self.assertAddsMessages(pylint.testutils.Message(msg_id='os-open-unsafe-permissions', node=node)):
+            with self.assertAddsMessages(
+                pylint.testutils.Message(
+                    msg_id='os-open-unsafe-permissions', node=node, args=(getattr(self.checker, '_os_open_msg_arg'),)
+                )
+            ):
                 self.checker.visit_call(node)
         else:
             with self.assertNoMessages():
@@ -181,7 +190,7 @@ class TestSecureCodingStandardChecker(pylint.testutils.CheckerTestCase):
     )
     def test_os_open_with_default_modes(self, mode, call_mode, arg, expected_warning):
         node = astroid.extract_node(call_mode.format(mode))
-        self.checker.set_os_open_mode(arg)
+        self.checker.set_os_open_allowed_modes(arg)
         if expected_warning:
             with self.assertAddsMessages(pylint.testutils.Message(msg_id='os-open-unsafe-permissions', node=node)):
                 self.checker.visit_with(node)
@@ -201,7 +210,7 @@ class TestSecureCodingStandardChecker(pylint.testutils.CheckerTestCase):
     )
     def test_os_open_with(self, mode, arg, expected_warning):
         node = astroid.extract_node(f'with os.open("file.txt", os.O_WRONLY, 0o{mode:o}) as fd: fd.read() #@')
-        self.checker.set_os_open_mode(arg)
+        self.checker.set_os_open_allowed_modes(arg)
         if expected_warning and mode != 0o755:
             with self.assertAddsMessages(pylint.testutils.Message(msg_id='os-open-unsafe-permissions', node=node)):
                 self.checker.visit_with(node)